backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
curl==0.0.1
Django==3.0.3
django-debug-toolbar==2.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
pathspec==0.7.0
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
regex==2020.2.20
requests==2.23.0
siaskynet==0.0.5
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
